refactor(engine): move supporting_factors to engine/ top-level (cross-approach)
supporting_factors.py lived under engine/sa/ but is consumed by all three
approach branches (SA, IRB, slotting) plus the pipeline orchestrator's
pre-split E* aggregation. Filing it under sa/ implied SA was a dependency
of irb/ and slotting/, when in fact supporting factors are cross-cutting.

Relocates the module to engine/supporting_factors.py as a peer of ccf.py,
hierarchy.py, classifier.py, re_splitter.py, and materialise.py. Pure
move: no behavioural change, no regulatory logic touched. git mv used to
preserve history.

Drops the dead re-export in engine/sa/__init__.py (no caller used it).
Updates LOGGER_REQUIRED_EXEMPT in scripts/arch_check.py and regenerates
docs/development/citation-matrix.md.

Verification: full dev-loop suite green (6804 passed, 2 skipped); ty
clean; arch_check clean for affected paths.
