+++ docs/appendix/changelog.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **On-balance-sheet netting now restricted to drawn loan siblings only (CRR Art. 219)**: `generate_netting_collateral` in `src/rwa_calc/engine/crm/collateral.py` was allocating synthetic cash collateral pro-rata across **every** positive-EAD sibling under the same netting facility — including off-balance-sheet contingents and synthetic `facility_undrawn` rows (the per-facility undrawn-headroom rows emitted by `hierarchy.py`). CRR Art. 219 explicitly limits OBS netting to drawn loans/deposits ("loans to and deposits with the lending institution"); the bank holds no irrevocable lending commitment on an off-BS row that a deposit could net against. Two coupled fixes: (a) the `positive_siblings` filter now requires `exposure_type == "loan"` AND `on_bs_for_ead > 0`, excluding contingents and facility_undrawn; (b) pro-rata basis switched from `ead_for_crm` (= `on_bs_for_ead + nominal_after_provision`, the CCF=100% override per Art. 223(4)) to `on_bs_for_ead` (the drawn portion), so a partly-drawn loan with a large undrawn nominal no longer captures an inflated share of the netting pool. The Art. 223(4) override remains the basis for FCCM E* netting against external cash collateral — it is a **collateral-valuation** rule, not an OBS-netting allocation rule. Source filter (negative-drawn loans with `has_netting_agreement=True`) was already correctly drawn-only and unchanged. Downstream collateral pipeline behaviour preserved: synthetic row still flagged `beneficiary_type="loan"` with `beneficiary_reference=exposure_reference`, lands at the `direct` allocation level and does not re-spread across facility/counterparty pools. Pinned by 4 new tests in `tests/unit/crm/test_netting.py::TestNettingDrawnOnlyScope` — contingent-excluded, facility_undrawn-excluded, pro-rata-uses-drawn-not-ead_for_crm (load-bearing anti-degenerate: deposit 200 with LOAN_A drawn=400 / no off-BS and LOAN_B drawn=100 / off-BS=900; old code split 57.14 / 142.86 by `ead_for_crm`, new code correctly splits 160 / 40 by drawn portion), and mixed-facility e2e. All 23 netting tests + 5,192 unit + 1,633 acceptance/integration/contract tests pass. Graceful fallback added for direct unit-test callers of `_generate_netting_collateral` that omit `exposure_type` / `on_bs_for_ead` (production always supplies them via the hierarchy + `_compute_ead` upstream). Docs updated: `docs/user-guide/methodology/crm.md` On-Balance Sheet Netting section now states drawn-only scope and `on_bs_for_ead` pro-rata basis with a mixed-facility example; `docs/specifications/crr/credit-risk-mitigation.md` adds an Art. 219 callout note distinguishing OBS-netting allocation basis from the Art. 223(4) collateral-valuation override. COREP C07 column 0035 (`on_bs_netting_amount`) semantics unchanged. Ref: CRR Art. 195, 205, 206, 219, 223(4), 224, 230; PRA PS1/26 Art. 219 (unchanged); BCBS CRE22.68-69.
 
 ### Changed
-- (Next release changes will go here)
+- **Supporting factors module relocated from `engine/sa/` to `engine/` top-level (cross-approach stage module)**: `src/rwa_calc/engine/sa/supporting_factors.py` → `src/rwa_calc/engine/supporting_factors.py`. The file housed both `SupportingFactorCalculator` (called by all three approach branches — SA at `engine/sa/namespace.py:2104`, IRB at `engine/irb/calculator.py:288`, slotting at `engine/slotting/calculator.py:166`) and the module-level helper `compute_e_star_group_drawn` (called by the pipeline orchestrator at `engine/pipeline.py:810` on the unified post-CRM frame before the SA/IRB/slotting split). Filing it under `engine/sa/` was a false-locality signal — `engine/irb/` and `engine/slotting/` both had to import from `engine/sa/`, implying SA was a dependency of the other approach branches when in fact supporting factors are cross-cutting. The module now sits as a peer of `engine/ccf.py`, `engine/hierarchy.py`, `engine/classifier.py`, `engine/re_splitter.py`, and `engine/materialise.py` — the other cross-cutting stage modules at the same level. Pure relocation: no behavioural change, no regulatory logic touched, no schema or output changes. Move performed via `git mv` to preserve history. Import paths updated in 5 source files (`engine/pipeline.py`, `engine/sa/namespace.py`, `engine/irb/calculator.py`, `engine/slotting/calculator.py`, `engine/sa/__init__.py`) and 4 test files (`tests/unit/test_supporting_factors.py`, `tests/unit/test_supporting_factors_cross_approach.py`, `tests/unit/crr/test_crr_sa.py`, `tests/contracts/test_watchfire_coverage.py`). The dead re-export in `engine/sa/__init__.py` (`SupportingFactorCalculator`, `create_supporting_factor_calculator`) was dropped since every caller already imported from the full module path. `LOGGER_REQUIRED_EXEMPT` in `scripts/arch_check.py` and the `--8<--` snippet path in `docs/user-guide/methodology/standardised-approach.md` updated to the new location. `docs/api/engine.md`, `docs/user-guide/methodology/supporting-factors.md`, `docs/data-model/output-schemas.md`, and `docs/specifications/common/default-definition.md` references updated; `docs/development/citation-matrix.md` regenerated via `scripts/generate_citation_matrix.py`. Historical changelog entries that mention the old path are intentionally left as-is. Full dev-loop suite green: 6804 passed, 2 skipped. Ref: CRR Art. 501 (SME), Art. 501a (infrastructure) — both regulations are themselves approach-agnostic, matching the new module location.
 
 ---
