+++ src/rwa_calc/engine/sa/namespace.py
@@ -158,7 +158,7 @@ from rwa_calc.data.tables.eu_sovereign import (
     denomination_currency_expr,
 )
 from rwa_calc.domain.enums import CQS, CRMCollateralMethod, EquityType
-from rwa_calc.engine.sa.supporting_factors import SupportingFactorCalculator
+from rwa_calc.engine.supporting_factors import SupportingFactorCalculator
 
 if TYPE_CHECKING:
     from rwa_calc.contracts.config import CalculationConfig
